留言模块实现

+++ src/main/java/site/liuming/truismcms/web/mapper/MessageMapper.java
@@ -38,5 +38,5 @@ public interface MessageMapper {
 
     List<Message> getList();
 
-    int replyMessage(@Param("id")Integer id);
+    int replyMessage(@Param("id")Long id);
 }
